requirements.txt
Flask
google-search-results
google-generativeai
supabase
python-dotenv
pandas
openpyxl
gunicorn
openai